refactor: 문자열 파싱을 담당하는 StringAddCalculatorParser 클래스 추가
